Remove analysis class and create submodule structure
