+++ asreview/state/legacy/dict.py
@@ -21,10 +21,46 @@ import numpy as np
 from scipy.sparse.csr import csr_matrix
 from scipy.sparse import save_npz, load_npz
 
-from asreview.analysis.statistics import _get_labeled_order
-from asreview.analysis.statistics import _get_last_proba_order
 from asreview.state.legacy.base import BaseState
 from asreview.settings import ASReviewSettings
+
+
+# from asreview.analysis.statistics DEPRECATED
+def _get_labeled_order(state):
+    """Get the order in which papers were labeled."""
+    label_order = []
+    n_initial = 0
+    n_predictor_models = state.n_predictor_models
+    for query_i in range(n_predictor_models):
+        try:
+            label_methods = state.get("label_methods", query_i)
+        except KeyError:
+            continue
+        label_idx = state.get("label_idx", query_i)
+        for i in range(len(label_idx)):
+            if label_methods[i] == "initial":
+                n_initial += 1
+        label_order.extend(label_idx)
+    return label_order, n_initial
+
+
+# from asreview.analysis.statistics DEPRECATED
+def _get_last_proba_order(state):
+    """Get the ranking of papers in the last query."""
+    n_predictor_models = state.n_predictor_models
+    pool_idx = None
+    for query_i in reversed(range(n_predictor_models)):
+        try:
+            pool_idx = state.get("pool_idx", query_i)
+        except KeyError:
+            continue
+        if pool_idx is not None:
+            proba = state.get("proba", query_i)
+            break
+
+    if pool_idx is None:
+        return []
+    return pool_idx[np.argsort(-proba[pool_idx])]
 
 
 def get_serial_list(array, dtype=None):
